+++ lib/archethic/self_repair.ex
@@ -11,13 +11,16 @@ defmodule Archethic.SelfRepair do
   alias __MODULE__.Scheduler
   alias __MODULE__.Sync
 
-  alias Archethic.BeaconChain
-
-  alias Archethic.Crypto
+  alias Archethic.{
+    BeaconChain,
+    Crypto,
+    Utils,
+    Contracts,
+    TransactionChain,
+    Election
+  }
 
   alias Archethic.P2P.Node
-
-  alias Archethic.Utils
 
   alias Crontab.CronExpression.Parser, as: CronParser
   alias Crontab.Scheduler, as: CronScheduler
@@ -158,5 +161,38 @@ defmodule Archethic.SelfRepair do
     changed_conf
     |> Keyword.get(Scheduler)
     |> Scheduler.config_change()
+  end
+
+  @doc """
+  Request missing transaction addresses from last local address until last chain address
+  and add them in the DB
+  """
+  def update_last_address(address, authorized_nodes) do
+    # As the node is storage node of this chain, it needs to know all the addresses of the chain until the last
+    # So we get the local last address and verify if it's the same as the last address of the chain
+    # by requesting the nodes which already know the last address
+
+    {last_local_address, _timestamp} = TransactionChain.get_last_address(address)
+    storage_nodes = Election.storage_nodes(last_local_address, authorized_nodes)
+
+    case TransactionChain.fetch_next_chain_addresses_remotely(last_local_address, storage_nodes) do
+      {:ok, []} ->
+        :ok
+
+      {:ok, addresses} ->
+        genesis_address = TransactionChain.get_genesis_address(address)
+
+        addresses
+        |> Enum.sort_by(fn {_address, timestamp} -> timestamp end)
+        |> Enum.each(fn {address, timestamp} ->
+          TransactionChain.register_last_address(genesis_address, address, timestamp)
+        end)
+
+        # Stop potential previous smart contract
+        Contracts.stop_contract(address)
+
+      _ ->
+        :ok
+    end
   end
 end
