+++ test/archethic/self_repair/repair_worker_test.exs
@@ -4,18 +4,11 @@ defmodule Archethic.SelfRepair.RepairWorkerTest do
 
   alias Archethic.BeaconChain.SummaryTimer
 
-  alias Archethic.Crypto
-
   alias Archethic.P2P
-  alias Archethic.P2P.Client.DefaultImpl
   alias Archethic.P2P.Node
-  alias Archethic.P2P.Message.GetNextAddresses
   alias Archethic.P2P.Message.GetTransaction
 
   alias Archethic.SelfRepair.RepairWorker
-
-  alias Archethic.TransactionChain.Transaction
-  alias Archethic.TransactionChain.Transaction.ValidationStamp
 
   import Mox
 
@@ -112,53 +105,5 @@ defmodule Archethic.SelfRepair.RepairWorkerTest do
              io_addresses: ["Bob1", "Bob2", "Bob3", "Bob4"],
              task: _task_pid
            } = :sys.get_state(pid)
-  end
-
-  test "update_last_address/1 should request missing addresses and add them in DB" do
-    node = %Node{
-      first_public_key: Crypto.first_node_public_key(),
-      last_public_key: Crypto.last_node_public_key(),
-      geo_patch: "AAA",
-      authorized?: true,
-      authorization_date: ~U[2022-11-27 00:00:00Z],
-      available?: true,
-      availability_history: <<1::1>>
-    }
-
-    me = self()
-
-    MockDB
-    |> expect(:get_last_chain_address, fn "Alice2" -> {"Alice2", ~U[2022-11-27 00:10:00Z]} end)
-    |> expect(:get_transaction, fn "Alice2", _ ->
-      {:ok, %Transaction{validation_stamp: %ValidationStamp{timestamp: ~U[2022-11-27 00:10:00Z]}}}
-    end)
-    |> expect(:get_first_chain_address, 2, fn "Alice2" -> "Alice0" end)
-    |> expect(:list_chain_addresses, fn "Alice0" ->
-      [
-        {"Alice1", ~U[2022-11-27 00:09:00Z]},
-        {"Alice2", ~U[2022-11-27 00:10:00Z]},
-        {"Alice3", ~U[2022-11-27 00:11:00Z]},
-        {"Alice4", ~U[2022-11-27 00:12:00Z]}
-      ]
-    end)
-    |> expect(:add_last_transaction_address, 2, fn
-      "Alice0", "Alice3", ~U[2022-11-27 00:11:00Z] ->
-        send(me, :add_alice3)
-
-      "Alice0", "Alice4", ~U[2022-11-27 00:12:00Z] ->
-        send(me, :add_alice4)
-    end)
-
-    MockClient
-    |> expect(:send_message, fn node, msg = %GetNextAddresses{address: "Alice2"}, timeout ->
-      send(me, :get_next_addresses)
-      DefaultImpl.send_message(node, msg, timeout)
-    end)
-
-    RepairWorker.update_last_address("Alice2", [node])
-
-    assert_receive :get_next_addresses
-    assert_receive :add_alice3
-    assert_receive :add_alice4
   end
 end
